requirements.txt
python-telegram-bot==21.2
python-dotenv==1.0.1
Flask==2.3.3
pytz==2023.3
gunicorn==21.2.0